Replace oteltest with sdk/trace/tracetest
