+++ instrumentation/net/http/otelhttp/test/config_test.go
@@ -24,16 +24,15 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
-	"go.opentelemetry.io/otel/oteltest"
+	"go.opentelemetry.io/otel/sdk/trace"
+	"go.opentelemetry.io/otel/sdk/trace/tracetest"
 )
 
 func TestBasicFilter(t *testing.T) {
 	rr := httptest.NewRecorder()
 
-	spanRecorder := new(oteltest.SpanRecorder)
-	provider := oteltest.NewTracerProvider(
-		oteltest.WithSpanRecorder(spanRecorder),
-	)
+	spanRecorder := tracetest.NewSpanRecorder()
+	provider := trace.NewTracerProvider(trace.WithSpanProcessor(spanRecorder))
 
 	h := otelhttp.NewHandler(
 		http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -58,7 +57,7 @@ func TestBasicFilter(t *testing.T) {
 	if got := rr.Header().Get("Traceparent"); got != "" {
 		t.Fatal("expected empty trace header")
 	}
-	if got, expected := len(spanRecorder.Completed()), 0; got != expected {
+	if got, expected := len(spanRecorder.Ended()), 0; got != expected {
 		t.Fatalf("got %d recorded spans, expected %d", got, expected)
 	}
 	d, err := ioutil.ReadAll(rr.Result().Body)
@@ -106,10 +105,8 @@ func TestSpanNameFormatter(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			rr := httptest.NewRecorder()
 
-			spanRecorder := new(oteltest.SpanRecorder)
-			provider := oteltest.NewTracerProvider(
-				oteltest.WithSpanRecorder(spanRecorder),
-			)
+			spanRecorder := tracetest.NewSpanRecorder()
+			provider := trace.NewTracerProvider(trace.WithSpanProcessor(spanRecorder))
 			handler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 				if _, err := io.WriteString(w, "hello world"); err != nil {
 					t.Fatal(err)
@@ -130,7 +127,7 @@ func TestSpanNameFormatter(t *testing.T) {
 				t.Fatalf("got %d, expected %d", got, expected)
 			}
 
-			spans := spanRecorder.Completed()
+			spans := spanRecorder.Ended()
 			if assert.Len(t, spans, 1) {
 				assert.Equal(t, tc.expected, spans[0].Name())
 			}
