+++ instrumentation/net/http/otelhttp/test/go.mod
@@ -5,9 +5,9 @@ go 1.15
 require (
 	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.22.0
-	go.opentelemetry.io/otel v1.0.0-RC2
+	go.opentelemetry.io/otel v1.0.0-RC2.0.20210812161231-a8bb0bf89f3b
 	go.opentelemetry.io/otel/metric v0.22.0
-	go.opentelemetry.io/otel/oteltest v1.0.0-RC2
+	go.opentelemetry.io/otel/sdk v1.0.0-RC2.0.20210812161231-a8bb0bf89f3b
 	go.opentelemetry.io/otel/trace v1.0.0-RC2
 )
